requirements.txt
asgiref==3.11.0
channels==4.3.2
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
sqlparse==0.5.5
dj-database-url
psycopg2-binary
